+++ main/filters.py
@@ -12,19 +12,8 @@
 """
 
 from jinja2 import Markup
-import re
 from datetime import datetime
-
-def parse_img_tags(mkd_text):
-    """
-    Replace every occurence of the <img src='...'> tag
-    with <img data-src='...'> tag for lazy loading images
-    in article view
-
-    ---- deprecated -------
-    """
-    return Markup(re.sub(r'<img .*src="',\
-     r'<img src="/static/src/img/ajax-loader.gif" data-src="', mkd_text))
+import re
 
 def timesince(dt, default="just now"):
     """