+++ main/models/UserProjectsModel.py
@@ -20,6 +20,10 @@ from main.helpers import handle_errors
 
 
 class UserProjects(BaseModel):
+    
+    """
+    Models and methods related to projects
+    """
 
     title = CharField(unique = True)
     body = TextField()
@@ -30,16 +34,12 @@ class UserProjects(BaseModel):
 
     @staticmethod
     def get_project(id):
-        try:
-            return UserProjects\
-                    .select()\
-                    .where(UserProjects.id == id)\
-                    .get()
-        except:
-            return 0
+        """ Get project by id """
+        return UserProjects.get_single("id", id)
 
     @staticmethod
     def get_all_projects():
+        """ Returns all projects or 0 if none exist"""
         q = UserProjects.select()
         if q.count():
             return q
@@ -48,6 +48,7 @@ class UserProjects(BaseModel):
 
     @staticmethod
     def check_exists(title, id = False):
+        """ Check if project with given title exists """
         try:
            q =  UserProjects.select().where((UserProjects.title == title))
            if not id:
@@ -55,9 +56,19 @@ class UserProjects(BaseModel):
            return q.where(UserProjects.id != id).get()
         except:
             return False
+
+
     @staticmethod
     @db.commit_on_success
-    def create_project(title, body, author):
+    def create_project(title, body, author, **kwargs):
+        """
+        Create new project
+        Arguments:
+            :title (unicode) - title of the project
+            :body (unicode) - body of the project
+            :author (object) - object containing author info
+
+        """
         if len(title) > 255:
             raise ValueError("Title must be at most 255 characters long")
         try:
@@ -73,6 +84,9 @@ class UserProjects(BaseModel):
     @staticmethod
     @db.commit_on_success
     def delete_project(project):
+        """
+        Delete project
+        """
         try:
             project.delete_instance()
             return 1
@@ -82,7 +96,16 @@ class UserProjects(BaseModel):
     
     @staticmethod
     @db.commit_on_success
-    def update_project(project, title, body):
+    def update_project(project, title, body, **kwargs):
+
+        """
+         Update project information
+         Arguments:
+             :project (object) - instance of project
+             :title (unicode) - new title
+             :body (unicode) - new body
+        """
+
         if len(title) > 255:
             raise ValueError("Title must be at most 255 characters long")
         try: