+++ main/views/base_views.py
@@ -21,6 +21,12 @@ import re
 
 class BaseView(MethodView):
 
+    
+    """
+    Base view to be implemented by every 
+    other class based view
+    """
+
     ARTICLE = False
 
     def get_template_name(self):
@@ -33,6 +39,7 @@ class BaseView(MethodView):
         with app.app_context():
             cache.clear()
         context.update(self.get_context())
+        context.update(dict(show_title = True))
         return render_template(self.get_template_name(), **context)
 
     def process_additional_fields(self):
@@ -47,6 +54,12 @@ class BaseView(MethodView):
 
 class ScratchpadView(BaseView):
 
+    
+    """
+    View to be implemented
+    by scratchpad views
+    """
+
     decorators = [login_required,]
 
     def get_template_name(self):
@@ -55,11 +68,25 @@ class ScratchpadView(BaseView):
 
 
 class ArticleView(BaseView):
+    
+    """
+    Additional view implemented
+    by update/create article views
+    """
 
     ARTICLE = True
 
 
     def process_additional_fields(self):
+
+        
+        """
+        Adds additional information needed for articles
+            : categories
+            : series
+            : article-image
+            : article-thumbnail
+        """
         categories = request.form.get("categories-hidden")
         categories = categories.strip()