+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 Tracking changes for Subrosa versions. Visit [https://github.com/exaroth/subrosa.git](https://github.com/exaroth/subrosa.git) for complete history revision.
 
+## 0.3
+
+* Complete dashboard panel rewrite
+* Added about page
+* Removed requests library from dependency list
+* Site title can now be customized in dashboard panel
+
 ## 0.2
 
 * Added categories and title image  to articles